Refactor get_from_dynamodb and save_in_dynamodb
