+++ api/make_private.py
@@ -6,10 +6,7 @@ import json
 from itertools import islice, product
 import decimal
 from shared.response import * 
-
-
-dynamodb = boto3.resource('dynamodb')
-table = dynamodb.Table("games")
+from shared.db import *
 
 
 def parse_event(event):
@@ -37,14 +34,6 @@ def is_valid_uuid(value):
         return True
     except ValueError:
         return False
-
-
-def get_from_dynamodb(game_uuid):
-    response = table.query(KeyConditionExpression=Key('game_uuid').eq(game_uuid))
-    items = response.get("Items")
-    if len(items) == 1:
-        return items[0]
-    return None
 
 
 def update_in_dynamodb(game_uuid, public):
